add production time to daughters in otf-decayer
