+++ ALICE3/Core/TrackUtilities.cxx
@@ -45,3 +45,71 @@ void o2::upgrade::convertTLorentzVectorToO2Track(const int charge,
   // Initialize TrackParCov in-place
   new (&o2track)(o2::track::TrackParCov)(x, particle.Phi(), params, covm);
 }
+
+float o2::upgrade::computeParticleVelocity(float momentum, float mass)
+{
+  const float a = momentum / mass;
+  // uses light speed in cm/ps so output is in those units
+  return o2::constants::physics::LightSpeedCm2PS * a / std::sqrt((1.f + a * a));
+}
+
+float o2::upgrade::computeTrackLength(o2::track::TrackParCov track, float radius, float magneticField)
+{
+  // don't make use of the track parametrization
+  float length = -100;
+
+  o2::math_utils::CircleXYf_t trcCircle;
+  float sna, csa;
+  track.getCircleParams(magneticField, trcCircle, sna, csa);
+
+  // distance between circle centers (one circle is at origin -> easy)
+  const float centerDistance = std::hypot(trcCircle.xC, trcCircle.yC);
+
+  // condition of circles touching - if not satisfied returned length will be -100
+  if (centerDistance < trcCircle.rC + radius && centerDistance > std::fabs(trcCircle.rC - radius)) {
+    length = 0.0f;
+
+    // base radical direction
+    const float ux = trcCircle.xC / centerDistance;
+    const float uy = trcCircle.yC / centerDistance;
+    // calculate perpendicular vector (normalized) for +/- displacement
+    const float vx = -uy;
+    const float vy = +ux;
+    // calculate coordinate for radical line
+    const float radical = (centerDistance * centerDistance - trcCircle.rC * trcCircle.rC + radius * radius) / (2.0f * centerDistance);
+    // calculate absolute displacement from center-to-center axis
+    const float displace = (0.5f / centerDistance) * std::sqrt(
+                                                       (-centerDistance + trcCircle.rC - radius) *
+                                                       (-centerDistance - trcCircle.rC + radius) *
+                                                       (-centerDistance + trcCircle.rC + radius) *
+                                                       (centerDistance + trcCircle.rC + radius));
+
+    // possible intercept points of track and TOF layer in 2D plane
+    const float point1[2] = {radical * ux + displace * vx, radical * uy + displace * vy};
+    const float point2[2] = {radical * ux - displace * vx, radical * uy - displace * vy};
+
+    // decide on correct intercept point
+    std::array<float, 3> mom;
+    track.getPxPyPzGlo(mom);
+    const float scalarProduct1 = point1[0] * mom[0] + point1[1] * mom[1];
+    const float scalarProduct2 = point2[0] * mom[0] + point2[1] * mom[1];
+
+    // get start point
+    std::array<float, 3> startPoint;
+    track.getXYZGlo(startPoint);
+
+    float cosAngle = -1000, modulus = -1000;
+
+    if (scalarProduct1 > scalarProduct2) {
+      modulus = std::hypot(point1[0] - trcCircle.xC, point1[1] - trcCircle.yC) * std::hypot(startPoint[0] - trcCircle.xC, startPoint[1] - trcCircle.yC);
+      cosAngle = (point1[0] - trcCircle.xC) * (startPoint[0] - trcCircle.xC) + (point1[1] - trcCircle.yC) * (startPoint[1] - trcCircle.yC);
+    } else {
+      modulus = std::hypot(point2[0] - trcCircle.xC, point2[1] - trcCircle.yC) * std::hypot(startPoint[0] - trcCircle.xC, startPoint[1] - trcCircle.yC);
+      cosAngle = (point2[0] - trcCircle.xC) * (startPoint[0] - trcCircle.xC) + (point2[1] - trcCircle.yC) * (startPoint[1] - trcCircle.yC);
+    }
+    cosAngle /= modulus;
+    length = trcCircle.rC * std::acos(cosAngle);
+    length *= std::sqrt(1.0f + track.getTgl() * track.getTgl());
+  }
+  return length;
+}
